consolidating views and components
