+++ packages/client/src/app/pages/my-stuff/components/content-item/content-item.component.less
