+++ packages/client/src/app/pages/my-stuff/components/content-item/content-item.component.ts
@@ -3,8 +3,8 @@ import { Select } from '@ngxs/store';
 import { ContentKind, ContentModel, PubStatus } from '@pulp-fiction/models/content';
 import { Observable } from 'rxjs';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { MyStuffState } from '../../../shared/my-stuff';
-import { MyStuffService } from '../my-stuff.service';
+import { MyStuffState } from '../../../../shared/my-stuff';
+import { MyStuffService } from '../../my-stuff.service';
 
 @UntilDestroy()
 @Component({
