+++ packages/client/src/app/pages/my-stuff/components/index.ts
@@ -1,1 +1,3 @@
 export { UploadCoverartComponent } from './upload-coverart/upload-coverart.component';
+export { ContentItemComponent } from './content-item/content-item.component';
+export { SectionItemComponent } from './section-item/section-item.component';
