+++ packages/client/src/app/pages/my-stuff/components/section-item/section-item.component.ts
@@ -5,9 +5,9 @@ import { Select } from '@ngxs/store';
 import { Observable } from 'rxjs';
 
 import { AuthorsNotePos, PublishSection, Section, SectionForm } from '@pulp-fiction/models/sections';
-import { SectionsState } from '../../../shared/my-stuff/sections';
-import { AlertsService } from '../../../shared/alerts';
-import { MyStuffService } from '../my-stuff.service';
+import { SectionsState } from '../../../../shared/my-stuff/sections';
+import { AlertsService } from '../../../../shared/alerts';
+import { MyStuffService } from '../../my-stuff.service';
 
 @UntilDestroy()
 @Component({
