+++ packages/client/src/app/pages/my-stuff/index.ts
@@ -1,14 +1,6 @@
 export { MyStuffComponent } from './my-stuff.component';
 export { MyStuffResolver } from './my-stuff.resolver';
-export { ContentItemComponent } from './content-item/content-item.component';
-export { ProseFormComponent } from './prose-form/prose-form.component';
-export { PoetryFormComponent } from './poetry-form/poetry-form.component';
-export { SectionItemComponent } from './section-item/section-item.component';
-
-export { BlogFormComponent } from './views/blog-form/blog-form.component';
-export { NewsFormComponent } from './views/news-form/news-form.component';
-export { ViewProseComponent } from './views/view-prose/view-prose.component';
-export { ViewPoetryComponent } from './views/view-poetry/view-poetry.component';
-export { ViewSectionsResolver } from './views/view-sections.resolver';
-
 export { MyStuffService } from './my-stuff.service';
+
+export * from './components';
+export * from './views';
