+++ packages/client/src/app/pages/my-stuff/views/index.ts
@@ -1,0 +1,7 @@
+export { ViewSectionsResolver } from './view-sections.resolver';
+export { BlogFormComponent } from './blog-form/blog-form.component';
+export { NewsFormComponent } from './news-form/news-form.component';
+export { PoetryFormComponent } from './poetry-form/poetry-form.component';
+export { ProseFormComponent } from './prose-form/prose-form.component';
+export { ViewPoetryComponent } from './view-poetry/view-poetry.component';
+export { ViewProseComponent } from './view-prose/view-prose.component';
