+++ packages/client/src/app/pages/my-stuff/views/prose-form/prose-form.component.ts
@@ -2,37 +2,34 @@ import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { Select } from '@ngxs/store';
 import { Observable } from 'rxjs';
-import { MyStuffState } from '../../../shared/my-stuff';
-
-import { WorkKind, CreatePoetry, PoetryForm, Genres, 
-  ContentRating, WorkStatus, PoetryContent, ContentKind } from '@pulp-fiction/models/content';
-import { AlertsService } from '../../../shared/alerts';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { MyStuffService } from '../my-stuff.service';
+import { MyStuffState } from '../../../../shared/my-stuff';
+
+import { WorkKind, Genres, ContentRating, WorkStatus, 
+    CreateProse, ProseContent, ContentKind } from '@pulp-fiction/models/content';
+import { AlertsService } from '../../../../shared/alerts';
+import { MyStuffService } from '../../my-stuff.service';
 
 @UntilDestroy()
 @Component({
-  selector: 'poetry-form',
-  templateUrl: './poetry-form.component.html',
-  styleUrls: ['./poetry-form.component.less']
+    selector: 'prose-form',
+    templateUrl: './prose-form.component.html',
+    styleUrls: ['./prose-form.component.less']
 })
-export class PoetryFormComponent implements OnInit {
-    @Select(MyStuffState.currContent) currContent$: Observable<PoetryContent>;
-    formTitle = `Create New Poetry`;
+export class ProseFormComponent implements OnInit {
+    @Select(MyStuffState.currContent) currContent$: Observable<ProseContent>;
+    formTitle = `Create New Prose`;
 
     categories = WorkKind;
-    forms = PoetryForm;
     genres = Genres;
     ratings = ContentRating;
     statuses = WorkStatus;
-    isCollection = false;
 
-    poetryForm = new FormGroup({
+    proseForm = new FormGroup({
         title: new FormControl('', [Validators.required, Validators.minLength(3), Validators.maxLength(100)]),
         desc: new FormControl('', [Validators.required, Validators.minLength(3), Validators.maxLength(250)]),
         body: new FormControl('', [Validators.required, Validators.minLength(3)]),
         category: new FormControl(null, [Validators.required]),
-        form: new FormControl(null, [Validators.required]),
         genres: new FormControl([], [Validators.required, Validators.minLength(1), Validators.maxLength(3)]),
         rating: new FormControl(null, [Validators.required]),
         status: new FormControl(null, [Validators.required])
@@ -44,25 +41,21 @@ export class PoetryFormComponent implements OnInit {
         this.currContent$.pipe(untilDestroyed(this)).subscribe(content => {
             if (content !== null) {
                 this.formTitle = `Editing "${content.title}"`;
-                this.isCollection = content.meta.collection;
-
-                this.poetryForm.setValue({
+                this.proseForm.setValue({
                     title: content.title,
                     desc: content.desc,
                     body: content.body,
                     category: content.meta.category,
-                    form: content.meta.form,
                     genres: content.meta.genres,
                     rating: content.meta.rating,
                     status: content.meta.status
                 });
             } else {
-                this.poetryForm.setValue({
+                this.proseForm.setValue({
                     title: '',
                     desc: '',
                     body: '',
                     category: null,
-                    form: null,
                     genres: [],
                     rating: null,
                     status: null
@@ -71,30 +64,28 @@ export class PoetryFormComponent implements OnInit {
         });
     }
 
-    get fields() { return this.poetryForm.controls; }
+    get fields() { return this.proseForm.controls; }
 
     submitForm(contentId?: string) {
-        if (this.poetryForm.invalid) {
+        if (this.proseForm.invalid) {
             this.alerts.warn(`Looks like something's wrong with the stuff you've entered.`);
             return;
         }
 
-        const poetryInfo: CreatePoetry ={
+        const proseInfo: CreateProse ={
             title: this.fields.title.value,
             desc: this.fields.desc.value,
             body: this.fields.body.value,
             category: this.fields.category.value,
-            collection: this.isCollection,
-            form: this.fields.form.value,
             genres: this.fields.genres.value,
             rating: this.fields.rating.value,
             status: this.fields.status.value
         };
 
         if (contentId) {
-            this.stuff.saveContent(contentId, ContentKind.PoetryContent, poetryInfo);
+            this.stuff.saveContent(contentId, ContentKind.ProseContent, proseInfo);
         } else {
-            this.stuff.createContent(ContentKind.PoetryContent, poetryInfo);
+            this.stuff.createContent(ContentKind.ProseContent, proseInfo);
         }
     }
 }
